StudentBootcampDetailController moved to new folder Student 	renamed:    app/Http/Controllers/api/StudentBootcampDetailController.php -> app/Http/Controllers/api/Student/StudentBootcampDetailController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Student/StudentBootcampDetailControllerTest.php
